Tests all run as the writer user now
The fix was that the data loaded by MLCP wasn't visible by the writer user. 

I think it'd be nice for some of these tests to eventually run as a "reader" user, but for now, this is a good improvement over having the tests all run as an "admin" user.