DEVEXP-257: Removed some debug logging
I was using GDS-DEBUG trace events while trying to build a GDS app, and I found all of these lines to either be verbose or useless. Probably a lot more that can be removed so it's easier to make sense of the logging when enabled, but this is a start.