DEVEXP-266: Updated license in source files
Note that 3 of the varianceplugin files were updated due to the text change in the .cpp file. I don't believe those files should be in git. But since they are regenerated as part of the user making use of `mlBundle`, it should not cause any harm for them to be in git.